// apps/game-client/src/lib/api/generated/main/model/update-user-preferences-dto-theme.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */

export type UpdateUserPreferencesDtoTheme = typeof UpdateUserPreferencesDtoTheme[keyof typeof UpdateUserPreferencesDtoTheme];


export const UpdateUserPreferencesDtoTheme = {
  default: 'default',
  cyberpunk: 'cyberpunk',
  pastel: 'pastel',
  fantasy: 'fantasy',
  shonen: 'shonen',
  onepiece: 'onepiece',
  anime: 'anime',
  goth: 'goth',
  halloween: 'halloween',
  realmadrid: 'realmadrid',
  'realmadrid-3rd': 'realmadrid-3rd',
  barcelona: 'barcelona',
  waguri: 'waguri',
  rukia: 'rukia',
  rias: 'rias',
  'cat-pink': 'cat-pink',
  'cat-purple': 'cat-purple',
  'cat-blue': 'cat-blue',
  'cat-random': 'cat-random',
} as const;