// apps/game-client/src/lib/api/generated/main/model/update-user-preferences-dto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */
import type { UpdateUserPreferencesDtoColorMode } from './update-user-preferences-dto-color-mode.js';
import type { UpdateUserPreferencesDtoMutes } from './update-user-preferences-dto-mutes.js';
import type { UpdateUserPreferencesDtoTheme } from './update-user-preferences-dto-theme.js';

export interface UpdateUserPreferencesDto {
  /** @minItems 1 */
  guildsOrder?: string[];
  theme?: UpdateUserPreferencesDtoTheme;
  colorMode?: UpdateUserPreferencesDtoColorMode;
  mutes?: UpdateUserPreferencesDtoMutes;
}