// apps/game-client/src/lib/api/generated/main/model/user-preferences-response-dto-output.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */
import type { UserPreferencesResponseDtoOutputMutes } from './user-preferences-response-dto-output-mutes';

export interface UserPreferencesResponseDtoOutput {
  userId: string;
  guildsOrder: string[];
  theme: string;
  colorMode: string;
  mutes: UserPreferencesResponseDtoOutputMutes;
}